refactor [DEV-27]: Rename the condition interfaces folder
