+++ libs/angular/ui/src/modules/condition/index.ts
@@ -1,3 +1,3 @@
 export { ConditionKeywordDirective } from './directives';
 export { CONDITION_KEYWORD } from './tokens';
-export { ConditionKeyword } from './models';
+export { ConditionKeyword } from './interfaces';
