+++ libs/angular/ui/src/modules/condition/interfaces/condition-keyword.ts
