+++ libs/angular/ui/src/modules/condition/interfaces/index.ts
