+++ libs/angular/ui/src/modules/condition/tokens/condition-keyword.token.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
 
-import { ConditionKeyword } from '../models';
+import { ConditionKeyword } from '../interfaces';
 
 export const CONDITION_KEYWORD = new InjectionToken<ConditionKeyword>('CONDITION_KEYWORD');
